+++ beecube-module-paimai/beecube-module-paimai-biz/src/main/java/cn/winkt/modules/paimai/service/im/message/GoodsUpdateMessage.java
@@ -4,6 +4,7 @@ package cn.winkt.modules.paimai.service.im.message;
 import com.alibaba.fastjson.annotation.JSONField;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -17,6 +18,6 @@ public class GoodsUpdateMessage extends BaseMessage{
     private Date endTime;
     @JSONField(format = "yyyy-MM-dd HH:mm:ss")
     private Date actualEndTime;
-    private Float dealPrice;
+    private BigDecimal dealPrice;
     private String dealUserId;
 }
